typescript 7, MediaDetailRender
